+++ packages/studio-base/src/players/buildRosbag2Player.ts
@@ -2,14 +2,14 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { rootGetDataProvider } from "@foxglove/studio-base/dataProviders/rootGetDataProvider";
-import { DataProviderDescriptor } from "@foxglove/studio-base/dataProviders/types";
 import RandomAccessPlayer from "@foxglove/studio-base/players/RandomAccessPlayer";
 import AutomatedRunPlayer from "@foxglove/studio-base/players/automatedRun/AutomatedRunPlayer";
 import PerformanceMeasuringClient from "@foxglove/studio-base/players/automatedRun/PerformanceMeasuringClient";
 import videoRecordingClient from "@foxglove/studio-base/players/automatedRun/videoRecordingClient";
 import { BuildPlayerOptions } from "@foxglove/studio-base/players/buildPlayer";
 import { Player } from "@foxglove/studio-base/players/types";
+import { rootGetDataProvider } from "@foxglove/studio-base/randomAccessDataProviders/rootGetDataProvider";
+import { RandomAccessDataProviderDescriptor } from "@foxglove/studio-base/randomAccessDataProviders/types";
 import {
   inVideoRecordingMode,
   inPlaybackPerformanceMeasuringMode,
@@ -17,7 +17,7 @@ import {
 import { getSeekToTime } from "@foxglove/studio-base/util/time";
 
 export function buildRosbag2PlayerFromDescriptor(
-  rootDescriptor: DataProviderDescriptor,
+  rootDescriptor: RandomAccessDataProviderDescriptor,
   options: BuildPlayerOptions,
 ): Player {
   if (inVideoRecordingMode()) {
