+++ packages/studio-base/src/randomAccessDataProviders/rootGetDataProvider.ts
@@ -19,6 +19,7 @@ import CombinedDataProvider from "@foxglove/studio-base/randomAccessDataProvider
 import MemoryCacheDataProvider from "@foxglove/studio-base/randomAccessDataProviders/MemoryCacheDataProvider";
 import ParseMessagesDataProvider from "@foxglove/studio-base/randomAccessDataProviders/ParseMessagesDataProvider";
 import RenameDataProvider from "@foxglove/studio-base/randomAccessDataProviders/RenameDataProvider";
+import Rosbag2DataProvider from "@foxglove/studio-base/randomAccessDataProviders/Rosbag2DataProvider";
 import WorkerDataProvider from "@foxglove/studio-base/randomAccessDataProviders/WorkerDataProvider";
 import createGetDataProvider from "@foxglove/studio-base/randomAccessDataProviders/createGetDataProvider";
 import {
@@ -33,6 +34,7 @@ const getDataProviderBase = createGetDataProvider({
   MemoryCacheDataProvider,
   ParseMessagesDataProvider,
   RenameDataProvider,
+  Rosbag2DataProvider,
   WorkerDataProvider,
 });
