+++ packages/studio-base/src/randomAccessDataProviders/standardDataProviderDescriptors.ts
@@ -44,3 +44,13 @@ export function getRemoteBagDescriptor(
 
   return wrapInWorker(bagDataProvider);
 }
+
+export function getLocalRosbag2Descriptor(
+  folder: FileSystemDirectoryHandle,
+): RandomAccessDataProviderDescriptor {
+  return {
+    name: CoreDataProviders.Rosbag2DataProvider,
+    args: { bagFolderPath: { type: "folder", folder } },
+    children: [],
+  };
+}
